chore: File name chart renamed to format-chart-percentage
