+++ src/utils/format-chart-percentage.ts
